+++ fe/src/components/Budget/BudgetView.tsx
@@ -4,12 +4,12 @@ import { useMemo, useState } from 'react';
 import AnalysisEmpty from '@/components/analysis/common/AnalysisEmpty';
 import AnalysisLoading from '@/components/analysis/common/AnalysisLoading';
 import AnalysisSection from '@/components/analysis/common/AnalysisSection';
-import BudgetRecommendDialog from '@/components/analysis/Budget/BudgetRecommendDialog';
-import BudgetSetupDialog from '@/components/analysis/Budget/BudgetSetupDialog';
+import BudgetRecommendDialog from '@/components/budget/BudgetRecommendDialog';
+import BudgetSetupDialog from '@/components/budget/BudgetSetupDialog';
 import { Button } from '@/components/ui/button';
 import { CalculatedBudgetItem } from '@/types/budgetGuide';
 import CategoryBudgetSetting from '@/components/analysis/CategoryBudgetSetting';
-import ConfirmDialog from '@/components/analysis/common/ConfirmDialog';
+import ConfirmDialog from '@/components/budget/ConfirmDialog';
 import { Progress } from '@/components/ui/progress';
 import ResponsivePanel from '@/components/panel/ResponsivePanel';
 import { Separator } from '@/components/ui/separator';
