+++ fe/src/components/Budget/steps/TemplateSelectionStep.tsx
@@ -2,7 +2,7 @@ import { CheckCircle2, Info } from 'lucide-react';
 
 import { BUDGET_TEMPLATES } from '@/constants/budgetTemplates';
 import { Card } from '@/components/ui/card';
-import DialogStepHeader from '@/components/analysis/Budget/common/DialogStepHeader';
+import DialogStepHeader from '@/components/budget/steps/DialogStepHeader';
 import { TemplateId } from '@/types/budgetGuide';
 import { cn } from '@/lib/utils';
